Bluft

+++ app/src/main/java/us/huseli/kiddo/data/types/interfaces/IListItemFile.kt
@@ -7,7 +7,7 @@ interface IListItemFile : IListItemBase {
     val filetype: FileType
     val lastmodified: String?
     val mimetype: String?
-    val size: Int?
+    val size: Long?
 
     @Suppress("unused")
     enum class FileType {
